+++ src/GauntletProjection.hpp
@@ -1,0 +1,7 @@
+#include <vector>
+
+class GauntletProjection {
+public:
+    constexpr static int gauntlets = 54;
+    static std::vector<bool> projectedIDs;
+};